+++ src/Microsoft.Azure.SignalR.Common/ServiceConnections/ClientConnectionBase.cs
@@ -1,0 +1,8 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.Azure.SignalR;
+
+internal abstract class ClientConnectionBase
+{
+}
